Moved running DAGMC OpenMC to its own example
