+++ .github/workflows/examples.yml
@@ -50,4 +50,5 @@ jobs:
             -e 'equilibria/fem_fixed_boundary' \
             -e 'codes/ext_code_script' \
             -e 'mesh/mesh_tutorial' \
+            -e 'radiation_transport/run_cad_neutronics' \
             ${RUN_DAGMC_EX} \
